make small change to load teh language environments more easily
